+++ modules/CLIModules/editConfigFile.js
@@ -2,12 +2,13 @@
 const fs = require('fs');
 const path = require('path');
 const writeConfigFile = require('../utils').writeJSONFile;
+const readConfigFile = require('../utils').readConfigFile;
 const style = require('../consoleStyling');
 const newTimestamp = require('../newTimestamp').small;
 
 module.exports = {
-  editConfigKey: (configPath, key, value, Callback) => {
-    fs.readFile(configPath, (err, data) => {
+  editConfigKey: (key, value, Callback) => {
+    readConfigFile((err, configFilePath, data) => {
       if (err)
         return console.error(err);
 
@@ -19,7 +20,7 @@ module.exports = {
         return console.error(style.styledError, e);
       }
 
-      writeConfigFile(path.dirname(configPath), 'merger-config', userConfig, (err, data) => {
+      writeConfigFile(path.dirname(configFilePath), 'merger-config', userConfig, (err, data) => {
         if (err)
           return console.error(err);
 
@@ -32,8 +33,8 @@ module.exports = {
     });
   },
 
-  addFileToConfig: (configPath, newSourceFile, Callback) => {
-    fs.readFile(configPath, 'utf-8', (err, data) => {
+  addFileToConfig: (newSourceFile, Callback) => {
+    readConfigFile((err, configFilePath, data) => {
       if (err)
         return console.error(style.styledError, err);
 
@@ -45,17 +46,53 @@ module.exports = {
         return console.error(style.styledError, e);
       }
 
-      writeConfigFile(path.dirname(configPath), 'merger-config', userConfig, (err, data) => {
+      writeConfigFile(path.dirname(configFilePath), 'merger-config', userConfig, (err, data) => {
         if (err)
           return console.error(err);
 
         let timestamp = newTimestamp();
-        console.info(`\n ${timestamp} - ${style.successText('Successsfuly added the new source file to the MergerJS configuration file successful.')}\n`, data);
+        console.info(`\n ${timestamp} - ${style.successText('Successsfuly added the new source file to the MergerJS configuration file.')}\n`, data);
 
         if (Callback)
           Callback();
       });
     });
+  },
 
+  removeSourceFile: (sourceFileObject, Callback) => {
+    readConfigFile((err, configFilePath, data) => {
+      if (err)
+        return console.error(style.styledError, err);
+
+      let userConfig
+      try {
+        userConfig = JSON.parse(data);
+        // const fileIndex = userConfig.sourceFiles.indexOf(sourceFileObject);
+        const searchFileIndex = (userConfig) => {
+          for (let i = 0; i < userConfig.sourceFiles.length; i++) {
+            if (JSON.stringify(userConfig.sourceFiles[i]) === JSON.stringify(sourceFileObject)) {
+              return i;
+            }
+          }
+          return -1;
+        };
+
+        const fileIndex = searchFileIndex(userConfig);
+
+        userConfig.sourceFiles.splice(fileIndex, 1);
+      } catch (e) {
+        return console.error(style.styledError, e);
+      }
+
+      writeConfigFile(path.dirname(configFilePath), 'merger-config', userConfig, (err, data) => {
+        if (err) return console.error(err);
+
+        let timestamp = newTimestamp();
+        console.info(`\n ${timestamp} - ${style.successText('Successsfuly removed the source file from the MergerJS configuration file.')}\n`, data);
+
+        if (Callback)
+          Callback();
+      });
+    });
   }
 }