+++ modules/checkForUpdates.js
@@ -1,6 +1,8 @@
-﻿const https = require('https');
+﻿'use strict'
+const https = require('https');
 const notify = require('./notifications').notif;
 const style = require('./consoleStyling');
+const UPDATE_ERROR = `\n ${style.warningTitle('Warn: ')} ${style.warningText('There was an error updating MergerJS.\n Please, check for updates manualy and if the problem persists, create an issue on GitHub.')}`;
 
 module.exports = (Callback) => {
   https.get('https://skimdb.npmjs.com/registry/merger-js', (res) => {
@@ -17,10 +19,12 @@ module.exports = (Callback) => {
         const latestVersion = parsedData['dist-tags'].latest;
 
         if (global.version !== latestVersion) {
+          // Different version.
           console.warn(`\n ${style.warningTitle('There is a newer version of MergerJS')} \n Please, run "npm i merger-js -g" or "merger update" to update.\n ${global.version} -> ${latestVersion }\n\n CHANGELOG: https://github.com/joao-neves95/merger-js/blob/master/CHANGELOG.md \n`);
           notify('New version of MergerJS available.', `Please, run "merger update" to update.\n${global.version} -> ${latestVersion}`);
           return Callback();
         } else {
+          // Current version.
           return Callback();
         }
       } catch (e) {
@@ -28,6 +32,7 @@ module.exports = (Callback) => {
       }
     })
   }).on('error', () => {
+    console.warn(UPDATE_ERROR);
     return Callback();
   });
 }